# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analyst/orchestration/runs:
    get:
      operationId: getAnalystOrchestrationRuns
      tags: [analyst]
      summary: List daily orchestration runs
      parameters:
        - name: date
          in: query
          required: false
          schema: { type: string, format: date }
      responses:
        "200":
          description: Orchestration history and schedule policy
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrchestrationRunList" }
  /analyst/orchestration/run:
    post:
      operationId: startAnalystOrchestrationRun
      tags: [analyst]
      summary: Start a manual daily orchestration run
      parameters:
        - name: date
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        "202":
          description: Run accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrchestrationRun" }
  /analyst/orchestration/runs/{runId}/interrupt:
    post:
      operationId: interruptAnalystOrchestrationRun
      tags: [analyst]
      summary: Request safe interruption of an active run
      parameters:
        - name: runId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Updated run record
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrchestrationRun" }
  /analyst/orchestration/late-scratches:
    post:
      operationId: detectAnalystLateScratches
      tags: [analyst]
      summary: Record post-freeze lineup scratches as immutable corrections
      parameters:
        - name: date
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        "201":
          description: Late-scratch detection result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LateScratchDetection" }
...
  /analyst/settlements/{gamePk}:
    post:
      operationId: settleOfficialGame
      tags: [analyst]
      summary: Settle one completed game from MLB Stats API
      parameters:
        - name: gamePk
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "201":
          description: Official game settlement result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfficialGameSettlementResult"
        "400":
          description: Invalid game or non-official settlement request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
                required: [error]
...
  /analyst/refresh/bullpen:
    post:
      operationId: refreshBullpen
      tags: [analyst]
      summary: Ingest reliever appearances for the last 3 days and recompute availability and leverage maps
      parameters:
        - name: date
          in: query
          required: false
          schema: { type: string, format: date }
      responses:
        "201":
          description: Bullpen ingest completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BullpenIngestResult"
components:
  schemas:
...
    BullpenIngestResult:
      type: object
      properties:
        source: { type: string }
        slateDate: { type: string }
        gamesProcessed: { type: integer }
        appearancesNormalized: { type: integer }
        appearancesRejected: { type: integer }
        teamsComputed: { type: integer }
        error: { type: ["string", "null"] }
      required: [source, slateDate, gamesProcessed, appearancesNormalized, appearancesRejected, teamsComputed, error]
...
    UnsettleableBoardCandidate:
      type: object
      properties:
        playerId: { type: integer }
        market: { type: string }
        reason: { type: string }
        gamePk: { type: integer }
      required: [playerId, market, reason]
    SettlementReconciliation:
      type: object
      description: |
        Every board candidate either settled or appears by name in unsettleable
        with the reason it did not. Settlement previously iterated only frozen
        feature snapshots, so a candidate that reached the board without a
        snapshot was never settled and vanished from the record silently.
      properties:
        boardCandidates: { type: integer, minimum: 0 }
        settled: { type: integer, minimum: 0 }
        settledWithoutSnapshot:
          type: integer
          minimum: 0
          description: |
            Settled from the board with no frozen snapshot. A complete
            operational settle record, deliberately excluded from model
            training because it has no pregame feature vector.
        unsettleable:
          type: array
          items: { $ref: "#/components/schemas/UnsettleableBoardCandidate" }
        totalBasesDiscrepancies:
          type: array
          description: |
            Rows where the total bases computed from the components and the
            total bases the feed reported disagree. Settled DISPUTED rather
            than silently taking one of the two.
          items: { type: string }
        walkDefinition:
          type: string
          description: The walk definition applied to this settlement, e.g. BB+IBB (assumed).
        walkDefinitionAssumed: { type: boolean }
        walkDefinitionStatement: { type: string }
      required: [boardCandidates, settled, settledWithoutSnapshot, unsettleable, totalBasesDiscrepancies, walkDefinition, walkDefinitionAssumed, walkDefinitionStatement]
    OfficialGameSettlementResult:
      type: object
      properties:
        gamePk: { type: integer, minimum: 1 }
        slateDate: { type: string, format: date }
        ingestRunId: { type: string }
        state: { type: string, enum: [PENDING, SETTLED, POSTPONED, NO_ACTION, DISPUTED] }
        lines: { type: integer, minimum: 0 }
        outcomesWritten: { type: integer, minimum: 0 }
        corrections: { type: integer, minimum: 0 }
        reconciliation: { $ref: "#/components/schemas/SettlementReconciliation" }
      required: [gamePk, slateDate, ingestRunId, state, lines, outcomesWritten, corrections]
...
    OrchestrationStep:
      type: object
      additionalProperties: false
      properties:
        name: { type: string }
        status: { type: string, enum: [PENDING, RUNNING, SUCCESS, WARNING, FAILED, CANCELLED] }
        startedAt: { type: ["string", "null"], format: date-time }
        finishedAt: { type: ["string", "null"], format: date-time }
        detail: { type: ["string", "null"] }
      required: [name, status, startedAt, finishedAt, detail]
    OrchestrationRun:
      type: object
      additionalProperties: false
      properties:
        runId: { type: string, format: uuid }
        runDate: { type: string, format: date }
        triggeredBy: { type: string, enum: [SCHEDULED, OPERATOR] }
        overallStatus: { type: string, enum: [RUNNING, COMPLETE, PARTIAL, FAILED, CANCELLED] }
        steps: { type: array, items: { $ref: "#/components/schemas/OrchestrationStep" } }
        schedule: { type: object, additionalProperties: true }
        frozenAt: { type: ["string", "null"], format: date-time }
        cancelRequestedAt: { type: ["string", "null"], format: date-time }
        errorMessage: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        finishedAt: { type: ["string", "null"], format: date-time }
      required: [runId, runDate, triggeredBy, overallStatus, steps, schedule, frozenAt, cancelRequestedAt, errorMessage, createdAt, finishedAt]
    OrchestrationRunList:
      type: object
      additionalProperties: false
      properties:
        runs: { type: array, items: { $ref: "#/components/schemas/OrchestrationRun" } }
        total: { type: integer }
        schedulePolicy: { type: object, additionalProperties: true }
      required: [runs, total, schedulePolicy]
    LateScratchDetection:
      type: object
      additionalProperties: false
      properties:
        slateDate: { type: string, format: date }
        corrections: { type: integer }
        targetedRerun: { type: boolean }
      required: [slateDate, corrections, targetedRerun]
...